Convert tests to PyTest
Very WIP conversion to use PyTest so tests can be parameterized tests.  New
structure reports tests more granularly, helping to pinpoint exact failure
cases. Unconverted tests are still in Unittest format, but should still run.

NOTE: 23.98 is failing the 24 hour wraparound test (and is marked as such). The
cause needs to be investigated and addressed so the xfail can be removed.

Found or addressed a number of things in the Timecode class:
* Timecode's init now takes time units instead of three possible starting values
with confusing override orders. Init now fails if both a timecode string and
time units are provided.
* Added safety and exception handling to add/sub_frames.
* Separated str/repr implementations; not sold on the current repr.
* Fixed str truncating third digit when framerate is milliseconds
* Removed old disabled tests